refactor: refine openvino
