+++ docs/openvino_requirements.txt
@@ -3,3 +3,4 @@ networkx==2.3.0
 numpy>=1.12.0
 test-generator==0.1.1
 defusedxml>=0.5.0
+serial
