+++ nn_meter/builder/backends/openvino/openvino_runner.py
@@ -1,22 +1,19 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 import os
-import tensorflow as tf
 import subprocess
 import numpy as np
 import shutil
+import serial
 
-from .utils import serial
-from .vpu_parser import VPUParser
-from ..interface import BaseBackend
-from .utils import keras_model_to_frozenpb
-from .frozenpb_patcher import patch_frozenpb
 from .utils import restart
-from nn_meter.utils.path import get_filename_without_ext
 from nn_meter.utils.pyutils import get_pyver
 
 
-class VPURunner:
+class OpenVINORunner:
+
+    device = None
+
     def __init__(self, venv, optimizer, runtime_dir, serial, graph_path='', output_dir='', data_type='FP16'):
         self._graph_path = graph_path
         self._venv = venv
@@ -64,7 +61,7 @@ class VPURunner:
                 f'{os.path.join(self._runtime_dir, "benchmark_app")} '
                 f'-i {input_path} '
                 f'-m {os.path.join(self._output_dir, filename + ".xml")} '
-                f'-d MYRIAD '
+                f'-d {self.device} '
                 f'-report_type detailed_counters '
                 f'-report_folder {self._output_dir} '
                 f'-niter 50 '
@@ -92,33 +89,3 @@ class VPURunner:
                     retry -= 1
 
         return output
-
-
-class VPUBackend(BaseBackend):
-    parser_class = VPUParser
-    runner_class = VPURunner
-
-    def get_params(self):
-        super().get_params()
-        self.runner_kwargs.update({
-            'venv': self.params['MOVIDIUS_ENV'],
-            'optimizer': self.params['OPTIMIZER_PATH'],
-            'runtime_dir': self.params['OPENVINO_RUNTIME_DIR'],
-            'serial': self.params['DEVICE_SERIAL'],
-            'data_type': self.params['DATA_TYPE'],
-        })
-        self.tmp_dir = self.params['TMP_DIR']
-        self.venv = self.params['MOVIDIUS_ENV']
-
-    def profile(self, model, model_name, shapes):
-        model_tmp_dir = os.path.join(self.tmp_dir, model_name)
-        pb_path, _ = keras_model_to_frozenpb(model, model_tmp_dir, model_name, shapes)
-        patched_pb_path = patch_frozenpb(pb_path, os.path.join(self.venv, 'bin/python'))
-        self.runner.load_graph(patched_pb_path, model_tmp_dir)
-        return self.parser.parse(self.runner.run(shapes)).latency
-
-    def profile_model_file(self, model_path, shapes):
-        model_name = get_filename_without_ext(model_path)
-        model = tf.keras.models.load_model(model_path)
-        return self.profile(model, model_name, shapes)
-
