+++ nn_meter/builder/backends/openvino/utils/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 from .converters import keras_model_to_frozenpb
-from .vpu_restart import restart 
+from .restart import restart
